analytics/requirements.txt
geopandas==0.14.1
googlemaps==4.10.0
hvplot==0.9.1
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4
pymongo==4.6.1
python-dotenv==1.0.0
scikit_learn==1.3.2
Shapely==2.0.2
tqdm==4.66.1
